Made an Audio Manager + button select sound
Audio Manager is meant to be carried into all scenes and holds an array of all the sounds we want in our game. Currently there is an issue with gameobjects losing their reference to Audio Manager when using an EventTrigger component.  It occurs if you start in the title screen, go to game, and then back to title screen. Once you are back in title screen, the reference to the Audio Manager game object is lost.